ピンのリファクタリング
